+++ llama-index-core/llama_index/core/prompts/default_prompts.py
@@ -362,26 +362,6 @@ DEFAULT_SIMPLE_INPUT_PROMPT = PromptTemplate(
 
 
 ############################################
-# Pandas
-############################################
-
-DEFAULT_PANDAS_TMPL = (
-    "You are working with a pandas dataframe in Python.\n"
-    "The name of the dataframe is `df`.\n"
-    "This is the result of `print(df.head())`:\n"
-    "{df_str}\n\n"
-    "Follow these instructions:\n"
-    "{instruction_str}\n"
-    "Query: {query_str}\n\n"
-    "Expression:"
-)
-
-DEFAULT_PANDAS_PROMPT = PromptTemplate(
-    DEFAULT_PANDAS_TMPL, prompt_type=PromptType.PANDAS
-)
-
-
-############################################
 # JSON Path
 ############################################
