+++ llama-index-experimental/llama_index/experimental/__init__.py
@@ -1,0 +1,6 @@
+from llama_index.experimental.query_engine.pandas.pandas_query_engine import (
+    PandasQueryEngine,
+)
+from llama_index.experimental.param_tuner.base import ParamTuner
+
+__all__ = ["PandasQueryEngine", "ParamTuner"]
