+++ llama-index-experimental/tests/BUILD
@@ -1,0 +1,1 @@
+python_tests()
